feat(config): fail if can't assign a community to a peering
When analyzing the priorities of a gateway for a VPC peering,
if a gateway is part of the group responsible for that peering,
fail if no community is available for its position in the
ranked list of gateways to service the peering. This can only
happen if the number of members in a group exceeds the numer of
communities available.

Also, simplify the logic, avoid unnecessary conversions between
integer types and improve the tests.

+++ config/src/converters/k8s/config/communities.rs
@@ -15,7 +15,7 @@ impl TryFrom<&GatewayAgentSpec> for PriorityCommunityTable {
             None => Ok(comtable),
             Some(map) => {
                 for (prio, community) in map {
-                    let priority: u32 = prio.parse().map_err(|e| {
+                    let priority: usize = prio.parse().map_err(|e| {
                         Self::Error::InvalidData(format!("community priority {prio}: {e}"))
                     })?;
                     comtable.insert(priority, community)?;
